commit on 6 jan

+++ Myka App/Controllers/Basket/basketModelClass/basketNewModelClass.swift
@@ -306,5 +306,5 @@ struct TotalNew: Codable {
 
 // MARK: - Chocdf
 struct Chocdf: Codable {
-    let quantity: Int?
+    let quantity: Double?
 }
